fix : BankCode 타입을 Long → Integer로 변경

+++ airbng-pay/src/main/java/com/airbng/pay/repository/BankInfoRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BankInfoRepository extends JpaRepository<BankInfo, Long> {
-    boolean existsByBankCode(Long bankCode);
+public interface BankInfoRepository extends JpaRepository<BankInfo, Integer> {
+    boolean existsByBankCode(Integer bankCode);
 
-    BankInfo findByBankCode(Long bankCode);
+    BankInfo findByBankCode(Integer bankCode);
 }
